Remove namespace

+++ include_db/TABLES.h
@@ -7,13 +7,13 @@ struct USERS_INFO {
     int user_id;
     string name;
     string surname;
-    int age;
+    uint8_t age;
     string gender;
     string faculty;
-    int course_number;
+    uint8_t course_number;
     string vk_link;
     string telegram_link;
-    int num_pairs;
+    uint32_t num_pairs;
     string description;
     USERS_INFO();
 };
